chore: pytest fixes and improvements (#265)
